requirements.txt
pandas==2.2.3
pyarrow==19.0.1
psycopg2==2.9.10
requests==2.32.3
python-dotenv==1.1.0
